requirements.txt
Pillow==10.2.0
tweepy==4.14.0